+++ packages/bluesnap-direct-integration/src/index.ts
@@ -2,3 +2,4 @@ export { default as BlueSnapDirectEcpPaymentMethod } from './BlueSnapDirectEcpPa
 export { default as BlueSnapDirectAlternativePaymentMethod } from './BlueSnapDirectAlternativePaymentMethod';
 export { default as BlueSnapDirectSepaPaymentMethod } from './BlueSnapDirectSepaPaymentMethod';
 export { default as BlueSnapDirectIdealPaymentMethod } from './BlueSnapDirectIdealPaymentMethod';
+export { default as BlueSnapV2PaymentMethod } from './BlueSnapV2PaymentMethod';