+++ packages/braintree-integration/src/BraintreeAcceleratedCheckout/BraintreeAcceleratedCheckoutPaymentMethod.test.tsx
@@ -44,8 +44,9 @@ describe('BraintreeAcceleratedCheckoutPaymentMethod', () => {
 
         expect(initializePayment).toHaveBeenCalledWith({
             methodId: props.method.id,
-            braintreeacceleratedcheckout: {
+            braintreefastlane: {
                 onInit: expect.any(Function),
+                onChange: expect.any(Function),
             },
         });
     });