+++ packages/credit-card-integration/src/CreditCardPaymentMethodComponent.tsx
@@ -61,7 +61,7 @@ interface CreditCardPaymentMethodDerivedProps {
     isPaymentDataRequired: boolean;
     shouldShowInstrumentFieldset: boolean;
     isInstrumentCardCodeRequired(instrument: Instrument, method: PaymentMethod): boolean;
-    isInstrumentCardNumberRequired(instrument: Instrument): boolean;
+    isInstrumentCardNumberRequired(instrument: Instrument, method: PaymentMethod): boolean;
     loadInstruments(): Promise<CheckoutSelectors>;
 }
 
@@ -206,7 +206,7 @@ class CreditCardPaymentMethodComponent extends Component<
         const shouldShowCreditCardFieldset = !shouldShowInstrumentFieldset || isAddingNewCard;
         const isLoading = isInitializing || isLoadingInstruments;
         const shouldShowNumberField = selectedInstrument
-            ? isInstrumentCardNumberRequiredProp(selectedInstrument)
+            ? isInstrumentCardNumberRequiredProp(selectedInstrument, method)
             : false;
         const shouldShowCardCodeField = selectedInstrument
             ? isInstrumentCardCodeRequiredProp(selectedInstrument, method)
@@ -314,7 +314,10 @@ class CreditCardPaymentMethodComponent extends Component<
                         selectedInstrument,
                         method,
                     ),
-                    isCardNumberRequired: isInstrumentCardNumberRequiredProp(selectedInstrument),
+                    isCardNumberRequired: isInstrumentCardNumberRequiredProp(
+                        selectedInstrument,
+                        method,
+                    ),
                     language,
                 })
             );