+++ packages/klarna-integration/src/klarnav2/KlarnaV2PaymentMethod.test.tsx
@@ -1,0 +1,143 @@
+import {
+    CheckoutSelectors,
+    CheckoutService,
+    createCheckoutService,
+    createLanguageService,
+    PaymentMethod,
+} from '@bigcommerce/checkout-sdk';
+import { mount } from 'enzyme';
+import { Formik } from 'formik';
+import { noop } from 'lodash';
+import React, { FunctionComponent } from 'react';
+
+import { HostedWidgetPaymentComponent } from '@bigcommerce/checkout/hosted-widget-integration';
+import {
+    createLocaleContext,
+    LocaleContext,
+    LocaleContextType,
+} from '@bigcommerce/checkout/locale';
+import {
+    CheckoutProvider,
+    PaymentMethodId,
+    PaymentMethodProps,
+} from '@bigcommerce/checkout/payment-integration-api';
+import {
+    getCheckout,
+    getCustomer,
+    getPaymentFormServiceMock,
+    getPaymentMethod,
+    getStoreConfig,
+} from '@bigcommerce/checkout/test-mocks';
+
+import KlarnaV2PaymentMethod from './KlarnaV2PaymentMethod';
+
+describe('when using Klarna payment', () => {
+    let method: PaymentMethod;
+    let checkoutService: CheckoutService;
+    let checkoutState: CheckoutSelectors;
+    let defaultProps: PaymentMethodProps;
+    let localeContext: LocaleContextType;
+
+    let PaymentMethodTest: FunctionComponent<PaymentMethodProps>;
+
+    beforeEach(() => {
+        checkoutService = createCheckoutService();
+        checkoutState = checkoutService.getState();
+        localeContext = createLocaleContext(getStoreConfig());
+        method = { ...getPaymentMethod(), id: 'pay_now', gateway: PaymentMethodId.Klarna };
+
+        jest.spyOn(checkoutState.data, 'getConfig').mockReturnValue(getStoreConfig());
+
+        jest.spyOn(checkoutService, 'deinitializePayment').mockResolvedValue(checkoutState);
+
+        jest.spyOn(checkoutService, 'initializePayment').mockResolvedValue(checkoutState);
+
+        jest.spyOn(checkoutState.data, 'getCheckout').mockReturnValue(getCheckout());
+
+        jest.spyOn(checkoutState.data, 'isPaymentDataRequired').mockReturnValue(true);
+
+        defaultProps = {
+            method,
+            checkoutService,
+            checkoutState,
+            paymentForm: getPaymentFormServiceMock(),
+            language: createLanguageService(),
+            onUnhandledError: jest.fn(),
+        };
+
+        PaymentMethodTest = (props) => (
+            <CheckoutProvider checkoutService={checkoutService}>
+                <LocaleContext.Provider value={localeContext}>
+                    <Formik initialValues={{}} onSubmit={noop}>
+                        <KlarnaV2PaymentMethod {...props} />
+                    </Formik>
+                </LocaleContext.Provider>
+            </CheckoutProvider>
+        );
+    });
+
+    afterEach(() => {
+        jest.clearAllMocks();
+    });
+
+    it('renders as hosted widget component', () => {
+        const container = mount(<PaymentMethodTest {...defaultProps} method={method} />);
+        const component = container.find(HostedWidgetPaymentComponent);
+
+        expect(component.props()).toEqual(
+            expect.objectContaining({
+                containerId: `${method.id}Widget`,
+                deinitializePayment: expect.any(Function),
+                initializePayment: expect.any(Function),
+                method,
+            }),
+        );
+    });
+
+    it('initializes method with required config when no instruments', () => {
+        jest.spyOn(checkoutState.data, 'getInstruments').mockReturnValue(undefined);
+
+        const container = mount(<PaymentMethodTest {...defaultProps} method={method} />);
+        const component = container.find(HostedWidgetPaymentComponent);
+
+        expect(component.props()).toEqual(
+            expect.objectContaining({
+                containerId: `${method.id}Widget`,
+                deinitializePayment: expect.any(Function),
+                initializePayment: expect.any(Function),
+                method,
+            }),
+        );
+    });
+
+    it('initializes method with required config when customer is defined', () => {
+        jest.spyOn(checkoutState.data, 'getCustomer').mockReturnValue(getCustomer());
+
+        const container = mount(<PaymentMethodTest {...defaultProps} method={method} />);
+        const component = container.find(HostedWidgetPaymentComponent);
+
+        expect(component.props()).toEqual(
+            expect.objectContaining({
+                containerId: `${method.id}Widget`,
+                deinitializePayment: expect.any(Function),
+                initializePayment: expect.any(Function),
+                method,
+            }),
+        );
+    });
+
+    it('initializes method with required config', () => {
+        mount(<PaymentMethodTest {...defaultProps} method={method} />);
+
+        expect(checkoutService.initializePayment).toHaveBeenCalledWith(
+            expect.objectContaining({
+                methodId: method.id,
+                gatewayId: method.gateway,
+                // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+                [`${method.gateway}v2`]: {
+                    container: `#${method.id}Widget`,
+                },
+            }),
+        );
+    });
+});